When the actor system gets an unhandled error, it shuts down. Unfortunately, this caused the process to exit with a zero exit status.
  https://groups.google.com/forum/#!topic/akka-user/xpO8UK7Y9HU

part of the problem is that the actor system is like a non-daemonic
Python thread.  It keeps the process alive, while it's running, but
doesn't affect the exit status.  Until I hear of a better solution, I
now wait for the system to exit and then call the Java aequivalent of
sys.exit.