Changed a wait_until to a sleep followed by an assertion.
wait_until is a bit weird in this case because the cache is growing
and shrinking so wait_until might fire on a transient, rather than a
stable point.

Based on some experiments (and observations in production), I think
there *is* still a cache bug and so I think we'll still see some
intermittent failures. I'll be curious to see them.